requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.9,<3.0
celery>=5.2,<6.0
redis>=4.0,<5.0
requests>=2.27,<3.0
stix2>=3.0,<4.0
taxii2-client>=2.3.0,<3.0
djangorestframework>=3.13,<4.0
python-dotenv>=0.19,<1.0
django-celery-results>=2.6.0
django-celery-beat>=2.8.0
